@@ -1,21 +1,21 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.jakewharton.butterknife'
 apply plugin: 'maven'
 apply plugin: 'maven-publish'
 apply plugin: 'com.jfrog.bintray'
+apply plugin: 'com.jakewharton.butterknife'
 
 group = "com.edwardstock.android"
-version = "1.0.0"
+version = "1.1.0"
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     buildToolsVersion buildTools
 
     resourcePrefix 'mp_'
 
     defaultConfig {
         minSdkVersion 18
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName version
 
@@ -50,14 +50,14 @@ afterEvaluate {
 }
 
 dependencies {
-    implementation "com.android.support:appcompat-v7:${libSupport}"
-    implementation "com.android.support:support-v4:${libSupport}"
-    implementation "com.android.support:recyclerview-v7:${libSupport}"
-    implementation "com.android.support:recyclerview-selection:28.0.0"
-    implementation "com.android.support:support-annotations:${libSupport}"
-    implementation "com.android.support:design:${libSupport}"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support.constraint:constraint-layout-solver:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview-selection:1.0.0'
+    implementation 'androidx.annotation:annotation:1.1.0'
+    implementation 'com.google.android.material:material:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.constraintlayout:constraintlayout-solver:1.1.3'
 
     implementation 'com.google.android.exoplayer:exoplayer:2.8.4'
 
@@ -68,18 +68,18 @@ dependencies {
     annotationProcessor "com.github.hotchemi:permissionsdispatcher-processor:3.3.1"
 
     annotationProcessor 'com.arello-mobile:moxy-compiler:1.5.3'
-    implementation 'com.jakewharton:butterknife:9.0.0-SNAPSHOT'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:9.0.0-SNAPSHOT'
+    implementation 'com.jakewharton:butterknife:10.2.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.1.0'
     implementation 'com.google.guava:guava:26.0-android'
-    implementation 'com.annimon:stream:1.2.0'
+    implementation 'com.annimon:stream:1.2.1'
     implementation 'com.jakewharton.timber:timber:4.7.1'
     implementation 'com.jsibbold:zoomage:1.2.0'
     implementation 'com.github.chrisbanes:PhotoView:2.1.4'
     implementation 'com.github.bumptech.glide:glide:4.8.0'
-    implementation 'com.google.code.gson:gson:2.8.5'
+    implementation 'com.google.code.gson:gson:2.8.6'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
     testImplementation 'junit:junit:4.12'
 }
 
 
-apply from: '../scripts/minter_publish.gradle'
+apply from: '../scripts/maven_publish.gradle'
