@@ -1,7 +1,7 @@
 package com.edwardstock.multipicker.internal;
 
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 public class MultiPickerFileProvider extends FileProvider {
 }
