@@ -6,9 +6,6 @@ import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.support.annotation.Nullable;
-import android.support.annotation.RestrictTo;
-import android.support.v4.content.FileProvider;
 
 import com.edwardstock.multipicker.PickerConfig;
 import com.edwardstock.multipicker.data.MediaFile;
@@ -18,6 +15,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
+import androidx.annotation.Nullable;
+import androidx.annotation.RestrictTo;
+import androidx.core.content.FileProvider;
 import timber.log.Timber;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
