@@ -1,8 +1,9 @@
 package com.edwardstock.multipicker.internal.widgets;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by hoanglam on 9/5/16.
