@@ -2,11 +2,11 @@ package com.edwardstock.multipicker.internal.widgets;
 
 import android.annotation.TargetApi;
 import android.content.Context;
-
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.widget.FrameLayout;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 /**
  * android-multipicker. 2018
