@@ -1,15 +1,15 @@
 package com.edwardstock.multipicker.internal.widgets;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.ViewGroup;
 
 import com.github.chrisbanes.photoview.PhotoView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.ViewPager;
 import timber.log.Timber;
 
 /**
