@@ -1,8 +1,6 @@
 package com.edwardstock.multipicker.picker.adapters;
 
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -17,6 +15,9 @@ import com.edwardstock.multipicker.data.Dir;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 
 public class DirsAdapter extends RecyclerView.Adapter<DirsAdapter.DirViewHolder> {
