@@ -2,10 +2,6 @@ package com.edwardstock.multipicker.picker.ui;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.Surface;
 import android.view.View;
@@ -21,6 +17,10 @@ import com.edwardstock.multipicker.data.MediaFile;
 import java.lang.ref.WeakReference;
 import java.util.Collections;
 
+import androidx.annotation.LayoutRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
