@@ -1,9 +1,9 @@
 package com.edwardstock.multipicker.picker.views;
 
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.RecyclerView;
-
 import com.edwardstock.multipicker.data.Dir;
+
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 
 /**
